+++ deploy-service/common/src/test/java/com/pinterest/deployservice/db/DBDAOTest.java
@@ -21,7 +21,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.pinterest.deployservice.bean.AcceptanceStatus;
 import com.pinterest.deployservice.bean.AgentBean;
 import com.pinterest.deployservice.bean.AgentErrorBean;
@@ -38,7 +37,6 @@ import com.pinterest.deployservice.bean.DeployState;
 import com.pinterest.deployservice.bean.DeployType;
 import com.pinterest.deployservice.bean.EnvironBean;
 import com.pinterest.deployservice.bean.GroupRolesBean;
-import com.pinterest.deployservice.bean.HostAgentBean;
 import com.pinterest.deployservice.bean.HostBean;
 import com.pinterest.deployservice.bean.HostState;
 import com.pinterest.deployservice.bean.HostTagBean;
@@ -64,7 +62,6 @@ import com.pinterest.deployservice.dao.DeployDAO;
 import com.pinterest.deployservice.dao.EnvironDAO;
 import com.pinterest.deployservice.dao.GroupDAO;
 import com.pinterest.deployservice.dao.GroupRolesDAO;
-import com.pinterest.deployservice.dao.HostAgentDAO;
 import com.pinterest.deployservice.dao.HostDAO;
 import com.pinterest.deployservice.dao.HostTagDAO;
 import com.pinterest.deployservice.dao.PromoteDAO;
@@ -84,7 +81,6 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.UUID;
 
 import org.apache.commons.dbcp.BasicDataSource;
 import org.apache.commons.lang3.builder.EqualsBuilder;
@@ -113,7 +109,6 @@ public class DBDAOTest {
     private static TagDAO tagDAO;
     private static ScheduleDAO scheduleDAO;
     private static UtilDAO utilDAO;
-    private static HostAgentDAO hostAgentDAO;
     private static BasicDataSource dataSource;
 
     @BeforeAll
@@ -138,7 +133,6 @@ public class DBDAOTest {
         tagDAO = new DBTagDAOImpl(dataSource);
         scheduleDAO = new DBScheduleDAOImpl(dataSource);
         utilDAO = new DBUtilDAOImpl(dataSource);
-        hostAgentDAO = new DBHostAgentDAOImpl(dataSource);
     }
 
     @AfterEach
@@ -1191,66 +1185,6 @@ public class DBDAOTest {
             assertNotNull(conn);
             utilDAO.releaseLock(lockName, conn);
         }
-    }
-
-    @Test
-    public void testHostAgentDAO() throws Exception {
-        final String hostId = "host-1";
-
-        // Test Insert and getById
-        HostAgentBean hostAgentBean = genDefaultHostAgentBean(hostId);
-        hostAgentDAO.insert(hostAgentBean);
-        HostAgentBean getByIdBean = hostAgentDAO.getHostById(hostId);
-        assertEquals(hostAgentBean, getByIdBean);
-
-        // Test Update and getById
-        hostAgentBean.setIp("192.168.0.1");
-        hostAgentDAO.update(hostId, hostAgentBean);
-        HostAgentBean getByIdBean2 = hostAgentDAO.getHostById(hostId);
-        assertEquals(hostAgentBean, getByIdBean2);
-
-        // Test getHostByName
-        HostAgentBean getByNameBean = hostAgentDAO.getHostByName(hostAgentBean.getHost_name());
-        assertEquals(hostAgentBean, getByNameBean);
-
-        // Test getDistinctHostsCount
-        long hostCount = hostAgentDAO.getDistinctHostsCount();
-        assertEquals(1, hostCount);
-
-        // Test getStaleHosts
-        List<HostAgentBean> staleHosts = hostAgentDAO.getStaleHosts(System.currentTimeMillis() - 100_000);
-        assertTrue(staleHosts.isEmpty());
-
-        List<HostAgentBean> staleHosts2 = hostAgentDAO.getStaleHosts(System.currentTimeMillis() + 100_000);
-        assertEquals(1, staleHosts2.size());
-        assertEquals(hostAgentBean, staleHosts2.get(0));
-
-        List<HostAgentBean> staleHosts3 = hostAgentDAO.getStaleHosts(
-            System.currentTimeMillis() - 100_000, System.currentTimeMillis() + 100_000);
-        assertEquals(1, staleHosts3.size());
-        assertEquals(hostAgentBean, staleHosts3.get(0));
-
-        // Test Delete
-        hostAgentDAO.delete(hostId);
-        HostAgentBean getByIdBean3 = hostAgentDAO.getHostById(hostId);
-        assertNull(getByIdBean3);
-        long hostCount2 = hostAgentDAO.getDistinctHostsCount();
-        assertEquals(0, hostCount2);
-
-    }
-
-    private HostAgentBean genDefaultHostAgentBean(String hostId) {
-        return HostAgentBean.builder()
-            .ip("127.0.0.1")
-            .host_id(hostId)
-            .host_name(UUID.randomUUID().toString())
-            .create_date(System.currentTimeMillis())
-            .last_update(System.currentTimeMillis())
-            .agent_version("1.0")
-            .auto_scaling_group("auto-scaling-group")
-            .normandie_status("normandie status normal")
-            .knox_status("knox status")
-            .build();
     }
 
     private EnvironBean genDefaultEnvBean(
